More consolidation 2
