+++ src/main/java/onelemonyboi/miniutilities/MiniUtilities.java
@@ -16,7 +16,7 @@ import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.InterModEnqueueEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import onelemonyboi.miniutilities.blocks.earth.EarthBlocks;
+import onelemonyboi.miniutilities.blocks.earth.GenericEarthBlock;
 import onelemonyboi.miniutilities.blocks.complexblocks.quantumquarry.QuantumQuarryBlock;
 import onelemonyboi.miniutilities.blocks.complexblocks.mechanicalminer.MechanicalMinerBlock;
 import onelemonyboi.miniutilities.blocks.complexblocks.mechanicalplacer.MechanicalPlacerBlock;
@@ -63,9 +63,9 @@ public class MiniUtilities {
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::setup);
         JSONLoader.loadJSON();
         EVENT_BUS.register(this);
-        EVENT_BUS.addListener(EarthBlocks::convertCursed);
-        EVENT_BUS.addListener(EarthBlocks::convertBlessed);
-        EVENT_BUS.addListener(EarthBlocks::convertBlursed);
+        EVENT_BUS.addListener(GenericEarthBlock::convertCursed);
+        EVENT_BUS.addListener(GenericEarthBlock::convertBlessed);
+        EVENT_BUS.addListener(GenericEarthBlock::convertBlursed);
         EVENT_BUS.addListener(UnstableShears::instantShear);
         EVENT_BUS.addListener(UnstableHoe::hoeTransformation);
         EVENT_BUS.addListener(WorldGen::generate);
